% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotelling-package.R
\docType{data}
\name{bottle.df}
\alias{bottle.df}
\title{Bottle data}
\description{
This data contains the elemental concentration of five different elements
(Manganese, Barium, Strontium, Zirconium, and Titanium) in samples of glass
taken from six different Heineken beer bottles. 20 measurements were taken
from each bottle.
}
\references{
R. L. Bennett. \emph{Aspects of the analysis and interpretation
of glass trace evidence}. Master's thesis, Department of Chemistry,
University of Waikato, 2002.
}
\keyword{datasets}
