% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotelling-package.R
\docType{data}
\name{container.df}
\alias{container.df}
\title{Container data}
\description{
This data contains the elemental concentration of nine different elements
(Titanium, Aluminium, Iron, Manganese, Magnesium, Calcium, Barium,
Strontium, and Zirconium) in specimens of glass taken from two different
containers. Ten measurements were taken from each container.
}
\references{
Jose R. Almirall. Discrimination of glass samples by solution
based ICP-OES PhD thesis, Department of Chemistry, Florida International
University, 1998.
}
\keyword{datasets}
