% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotelling-package.R
\docType{package}
\name{Hotelling-package}
\alias{Hotelling-package}
\alias{Hotelling}
\title{Hotelling
A set of R functions and data sets which implements Hotelling's T^2 test, and some variants of it. 
Functions are also included for Aitchison's additive log ratio and centred log ratio transformations.}
\description{
Hotelling
A set of R functions and data sets which implements Hotelling's T^2 test, and some variants of it. 
Functions are also included for Aitchison's additive log ratio and centred log ratio transformations.
}
\author{
James Curran Maintainer: James Curran <j.curran@auckland.ac.nz> ~~
The author and/or maintainer of the package ~~
}
\keyword{package}
