% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotelling-package.R
\docType{data}
\name{manova1.df}
\alias{manova1.df}
\title{manova1 data}
\description{
The data contains example data for testing the unequal variance option in the 
package. The dataset has four varibles, \code{wratr}, \code{wrata}, \code{treatment}, 
and \code{disability}. \code{treatment is the grouping variable} and \code{wratr} and \code{wrata}
are the responses. \code{disability} is not used.
}
\references{
\url{https://ncss-wpengine.netdna-ssl.com/wp-content/themes/ncss/pdf/Procedures/NCSS/Hotellings_Two-Sample_T2.pdf}
}
\keyword{datasets}
