% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise}
\alias{summarise}
\alias{summarize}
\alias{summarise.default}
\alias{summarise.formula}
\alias{summarise.data.frame}
\title{Summary statistics for grouped data}
\usage{
summarise(x, ...)

\method{summarise}{default}(
  x,
  y,
  stats = list(Mean = mean, Median = median, `Std. Dev.` = sd, N = length)
)

\method{summarise}{formula}(
  x,
  data = NULL,
  stats = list(Mean = mean, Median = median, `Std. Dev.` = sd, N = length),
  ...
)

\method{summarise}{data.frame}(x, y, ...)
}
\arguments{
\item{x}{a matrix of multivariate observations, a list of summary statistics from
multivariate observations, a data.frame of multivariate observations, or a formula
with a multivariate response on the left hand side, and a grouping variable/factor
on the right hand side.}

\item{\ldots}{other arguments such as another matrix of multivariate observations:
see \code{summarise.default}, or a data to be used with a formula: see
\code{summarise.formula}}
}
\description{
Easily get summary statistics for each group present in the data
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Summary statistics for grouped data

\item \code{formula}: Summary statistics for grouped data

\item \code{data.frame}: Summary statistics for grouped data
}}

\examples{

data(container.df)
split.data = split(container.df[,-1],container.df$gp)
x = split.data[[1]]
y = split.data[[2]]
summarise(x, y)

## Using the formula interface
data(container.df)
summarise(gp~., data = container.df)

summarise(gp~Al+Ti, data = container.df)

}
