\name{container.df}
\docType{data}
\alias{container.df}
\title{Container data}
\description{
This data contains the elemental concentration of nine different
elements (Titanium, Aluminium, Iron, Manganese, Magnesium, Calcium,
Barium, Strontium, and Zirconium) in
specimens of glass taken from two different containers. Ten measurements
were taken from each container.
}
\usage{data(container.df)}
\references{
Jose R. Almirall. Discrimination of glass samples by solution based
ICP-OES PhD thesis, Department of Chemistry, Florida International
University, 1998.
}
\keyword{datasets}